app/requirements.txt
azure-identity>=1.19.0
azure-keyvault-secrets>=4.9.0
azure-storage-blob>=12.24.0
azure-cosmos>=4.9.0
azure-search-documents>=11.6.0
azure-ai-projects>=1.0.0
aiohttp>=3.11.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
